% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{confoundvis-package}
\alias{confoundvis-package}
\alias{confoundvis}
\title{confoundvis: Geometric Visualization Tools for Causal Sensitivity Analysis}
\description{
Provides visualization tools for sensitivity analysis under unmeasured
confounding, including sensitivity contour plots, robustness curves,
sensitivity Love plots, and local Taylor diagnostic plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/causalfragility-lab/confoundvis}
  \item Report bugs at \url{https://github.com/causalfragility-lab/confoundvis/issues}
}

}
\author{
\strong{Maintainer}: Subir Hait \email{haitsubi@msu.edu} (\href{https://orcid.org/0009-0004-9871-9677}{ORCID})

}
\keyword{internal}
